major re-org of goprotoc code: split into multiple files and start addressing some TODOs
