+++ cmd/goprotoc/free_fields.go
@@ -1,1 +1,12 @@
 package main
+
+import (
+	"io"
+
+	"github.com/jhump/protoreflect/desc"
+)
+
+func doPrintFreeFieldNumbers(fds []*desc.FileDescriptor, w io.Writer) error {
+	// TODO
+	return nil
+}
